Move game related code into a game/ subfolder
This is to split engine-specific code that's more or less general from the, in this case, Tetris-specific game logic and such
