+++ client/src/game/Game.ts
@@ -1,5 +1,5 @@
-import { HasLifecycle, IHasLifecycle } from 'engine/behavior/HasLifecycle'
-import { Field } from 'engine/Field'
+import { HasLifecycle, IHasLifecycle } from '../engine/behavior/HasLifecycle'
+import { Field } from './Field'
 
 export class Game extends HasLifecycle implements IHasLifecycle {
   protected field: Field
