+++ client/src/game/shapes/Coordinate.ts
