+++ client/src/game/shapes/Shapes.ts
@@ -1,0 +1,17 @@
+import { JShape } from './definitions/JShape'
+import { LShape } from './definitions/LShape'
+import { TShape } from './definitions/TShape'
+import { ZShape } from './definitions/ZShape'
+import { SShape } from './definitions/SShape'
+import { IShape } from './definitions/IShape'
+import { OShape } from './definitions/OShape'
+
+export const Shapes = {
+  0: LShape,
+  1: JShape,
+  2: TShape,
+  3: ZShape,
+  4: SShape,
+  5: IShape,
+  6: OShape,
+}
