+++ client/src/game/shapes/definitions/IShape.ts
@@ -1,4 +1,4 @@
-import { AbstractShape } from 'shapes/AbstractShape'
+import { AbstractShape } from '../AbstractShape'
 
 export class IShape extends AbstractShape {
   definition = [
