+++ client/src/game/shapes/definitions/JShape.ts
@@ -1,4 +1,4 @@
-import { AbstractShape } from 'shapes/AbstractShape'
+import { AbstractShape } from '../AbstractShape'
 
 export class JShape extends AbstractShape {
   definition = [
