+++ client/src/game/shapes/definitions/LShape.ts
@@ -1,4 +1,4 @@
-import { AbstractShape } from 'shapes/AbstractShape'
+import { AbstractShape } from '../AbstractShape'
 
 export class LShape extends AbstractShape {
   definition = [
