+++ client/src/game/shapes/definitions/SShape.ts
@@ -1,4 +1,4 @@
-import { AbstractShape } from 'shapes/AbstractShape'
+import { AbstractShape } from '../AbstractShape'
 
 export class SShape extends AbstractShape {
   definition = [
