+++ client/src/game/shapes/definitions/TShape.ts
@@ -1,4 +1,4 @@
-import { AbstractShape } from 'shapes/AbstractShape'
+import { AbstractShape } from '../AbstractShape'
 
 export class TShape extends AbstractShape {
   definition = [
