+++ client/src/game/shapes/definitions/ZShape.ts
@@ -1,4 +1,4 @@
-import { AbstractShape } from 'shapes/AbstractShape'
+import { AbstractShape } from '../AbstractShape'
 
 export class ZShape extends AbstractShape {
   definition = [
