--- client/src/shapes/Shapes.ts
@@ -1,17 +1,0 @@
-import { JShape } from './JShape'
-import { LShape } from './LShape'
-import { TShape } from './TShape'
-import { ZShape } from './ZShape'
-import { SShape } from './SShape'
-import { IShape } from './IShape'
-import { OShape } from './OShape'
-
-export const Shapes = {
-  0: LShape,
-  1: JShape,
-  2: TShape,
-  3: ZShape,
-  4: SShape,
-  5: IShape,
-  6: OShape,
-}
